requirement.txt
streamlit
requests
beautifulsoup4
PyMuPDF
langchain
langchain-groq
python-dotenv
